feat(cobro): el producto no cobraba, y ahora sí — auditoría de flujos
Encargo: revisar los flujos, no la interfaz. La primera pregunta —«en el
onboarding, ¿dónde paga el usuario?»— tenía la peor respuesta posible: en
ninguna parte, nunca.

Registrarse creaba una cuenta `starter`, que es el plan que /pricing cobra a
$80.000/mes, sin suscripción, sin vencimiento y sin ninguna pantalla que
volviera a pedir dinero. `billing_status` lo escribía el webhook y no lo leía
nadie. El checkout de Polar estaba construido y correcto; nada mandaba a nadie
hacia él. Ninguna de las 290 pruebas verdes podía verlo: cada una comprobaba
una pieza que sí funcionaba. No era un fallo, era una ausencia.

MURO DE PAGO (mig 106)

`accounts.access_state` (pending | active | delinquent), y se EXTIENDE
`app.company_is_active` en vez de crear políticas: esa función es el predicado
de las 543 políticas RESTRICTIVE de la migración 99, así que enseñarle la
columna las enseñó a las 543 a la vez. El muro vive en la base de datos, no
solo en un redirect que se esquiva hablando con PostgREST.

La condición es «la cuenta está al día O la empresa sigue configurándose»: sin
esa segunda mitad el paso de sucursales fallaría para toda cuenta nueva y el
cliente vería el muro antes de saber qué compra. Todo lo preexistente queda en
`active` por grandfathering.

App: /suscripcion fuera del grupo (dashboard), redirect en el layout DESPUÉS
del asistente, paso «Plan» sin salidas gratis. El botón «Subir a {tier}» tenía
`plan: 'growth'` escrito dentro, así que en una cuenta Growth la tarjeta de
Starter cobraba Growth.

CADENA COMERCIAL

`invoices.sales_order_id` existía desde la migración 98 —con FK y guardia
anti-cruce— y ni un archivo del repositorio la nombraba. Nuevo `facturarPedido`,
que convierte el IVA con `netFromGross`: cotización y pedido guardan el precio
CON impuesto y la factura SIN él, así que copiar la línea tal cual cobra el 19%
dos veces (el error que encontró la mig 104, por otra puerta). `invoiceTotals`
sale a lib/domain.ts al tener dos llamadores.

NUEVE FUNCIONES QUE EXISTÍAN Y NO LLAMABA NADIE

setLoanStatus, setNonconformityAction, setPresupuestoValor, updateReserva,
updateContratoArriendo, updatePrograma, updateCiclo, updateMesa, updateBranding.
Un préstamo nacía activo y moría activo; una no conformidad no podía registrar
su acción correctiva; una reserva solo se corregía borrándola.

Y dos secciones condenadas a estar vacías: `employee_skills` y
`employee_events` tienen tabla desde la migración 02, la ficha del empleado las
leía, y no existía un solo insert en todo el repositorio.

DINERO QUE SE PERDÍA AL GUARDAR

Cotizaciones descartaba en silencio la línea con precio y sin descripción,
mientras el editor mostraba su total en vivo: se veía $250.000 y se guardaba $0,
y ese pedido nacía sin líneas. Mismo arreglo en compras.

HONESTIDAD

Notificaciones, Marketing y Suscripciones dicen ahora en pantalla qué hacen y
qué no (no hay cron ni proveedor de envío); Marketing gana la descarga de
destinatarios, que es lo que lo vuelve útil hoy. Cuatro afirmaciones falsas
corregidas, una de ellas en los Términos de servicio. Cartera prellena la deuda
desde la factura en vez de llevar una segunda contabilidad.

LOGO (migs 107, 108)

Bucket privado `logos` con permiso `configuracion:manage`, y `valid_branding`
pasa a admitir una ruta de storage además de una URL https — la regla original
es de la mig 30, de cuando la única forma imaginada de tener logo era alojarlo
fuera. Era el último control muerto. «Cambiar foto» ya funcionaba.

MIGRACIÓN 105

Nunca se había aplicado, y no por olvido: `place_storefront_order` cambia el
tipo de un parámetro OUT, y `create or replace` lo rechaza aunque la firma de
entrada sea idéntica. Añadido el `drop function` y la reconcesión de permisos
que el drop se lleva.

GUARDIAS NUEVAS

Un `export const` desde un archivo 'use server' tumba TODA la ruta en ejecución
y ni tsc ni el build lo ven — se descubrió con seis specs de e2e en rojo y una
traza que señalaba a otro archivo. Pineado en use-server-exports.test.ts, y
comprobado que falla al reintroducirlo. paywall.test.ts pinea el muro entero.

E2e 6 → 8: suscripcion.spec.ts prueba el muro incluida la guardia de la base;
logo.spec.ts va por navegador porque psql entra como postgres y no evalúa
políticas de storage; embudo.spec.ts llega hasta la factura y ahora mira
importes, que es lo que destapó las líneas descartadas.

tsc 0 · vitest 303/303 · build verde · e2e 8/8 · migs 105–108 aplicadas.

Queda fuera a propósito: la integración de visores @extend-ai que estaba en el
árbol antes de esta sesión, y de donde salen los 17 errores de lint.

+++ src/server/mutations/productos.ts
@@ -31,7 +31,7 @@ const baseSchema = z.object({
   // El precio ya lo incluye; esto solo declara cuánto de él es impuesto, para
   // que el POS lo pueda desglosar y la factura descontarlo. Ver migración 104.
   taxRate: z.number().min(0, 'La tasa no puede ser negativa.').max(100, 'La tasa no puede pasar de 100.').default(0),
-  stock: z.number().int().min(0, 'El stock no puede ser negativo.').default(0),
+  stock: z.number().min(0, 'El stock no puede ser negativo.').default(0),
   supplier: z.string().trim().max(160).default(''),
   isActive: z.boolean().default(true),
   inStorefront: z.boolean().default(true),
